requirements.txt
Flask
Authlib
SpeechRecognition
google-generativeai
python-dotenv
requests